% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countCellAdjacenciesCpp}
\alias{countCellAdjacenciesCpp}
\title{Count cell adjacencies in a grid (c++)}
\usage{
countCellAdjacenciesCpp(vals, nrow, ncol, doublecount, eightconn = FALSE)
}
\arguments{
\item{vals}{\link{numeric}\cr the values to analyse (row-major flat
vector)}

\item{nrow}{\link[=integer]{integer(1)}\cr number of rows in the grid}

\item{ncol}{\link[=integer]{integer(1)}\cr number of columns in the grid}

\item{doublecount}{\link[=logical]{logical(1)}\cr whether or not to tally up each
adjacency for both neighbouring cells, or only once.}

\item{eightconn}{\link[=logical]{logical(1)}\cr whether to use 8-connectivity
(queen: also count the four diagonal neighbours) or 4-connectivity (rook:
orthogonal neighbours only).}
}
\value{
A matrix of values*values with their adjacencies
}
\description{
C++ function that counts the adjacencies between unique values in a flat
grid vector
}
\seealso{
Other count functions: 
\code{\link{countCellEdgesCpp}()},
\code{\link{countCellValuesCpp}()}
}
\concept{count functions}
