% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countCellEdgesCpp}
\alias{countCellEdgesCpp}
\title{Count cell edges in a grid (c++)}
\usage{
countCellEdgesCpp(vals, nrow, ncol)
}
\arguments{
\item{vals}{\link{numeric}\cr the values to analyse (row-major flat
vector)}

\item{nrow}{\link[=integer]{integer(1)}\cr number of rows in the grid}

\item{ncol}{\link[=integer]{integer(1)}\cr number of columns in the grid}
}
\value{
A data.frame of the unique values and their edges in X and Y
}
\description{
C++ function that counts the number of edges between different values in a
flat grid vector
}
\seealso{
Other count functions: 
\code{\link{countCellAdjacenciesCpp}()},
\code{\link{countCellValuesCpp}()}
}
\concept{count functions}
