% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countCellValuesCpp}
\alias{countCellValuesCpp}
\title{Count unique cell values in a grid (c++)}
\usage{
countCellValuesCpp(vals, nrow, ncol)
}
\arguments{
\item{vals}{\link{numeric}\cr the values to count (column-major flat
vector)}

\item{nrow}{\link[=integer]{integer(1)}\cr number of rows in the grid}

\item{ncol}{\link[=integer]{integer(1)}\cr number of columns in the grid}
}
\value{
A data.frame of the unique values and their number
}
\description{
C++ function that counts the number of unique values in a flat grid vector
}
\seealso{
Other count functions: 
\code{\link{countCellAdjacenciesCpp}()},
\code{\link{countCellEdgesCpp}()}
}
\concept{count functions}
