% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distanceCpp}
\alias{distanceCpp}
\title{Calculate a distance map (c++)}
\usage{
distanceCpp(vals, nrow, ncol, method)
}
\arguments{
\item{vals}{\link[=numeric]{numeric(.)}\cr flat vector of cell values
(row-major order). Foreground cells have value 1.}

\item{nrow}{\link[=integer]{integer(1)}\cr number of rows in the grid.}

\item{ncol}{\link[=integer]{integer(1)}\cr number of columns in the grid.}

\item{method}{\link[=character]{character(1)}\cr which of the methods
\code{"euclidean"}, \code{"manhattan"} or \code{"chessboard"} to use.}
}
\value{
A numeric vector of the same length with distance values.
}
\description{
C++ function that derives a distance map of foreground patches in a flat
grid vector.
}
\concept{grid modify functions}
