% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isBinaryCpp}
\alias{isBinaryCpp}
\title{Check whether cell values are binary (c++)}
\usage{
isBinaryCpp(vals)
}
\arguments{
\item{vals}{\link[=numeric]{numeric(.)}\cr flat vector of cell values.}
}
\value{
A single logical value.
}
\description{
C++ function that checks if all non-NA values in a numeric vector are
either 0 or 1.
}
