% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_centroid.R
\name{mdf_centroid}
\alias{mdf_centroid}
\title{Determine the centroid of patches}
\usage{
mdf_centroid(obj = NULL, background = NA, layer = NULL, add = NULL)
}
\arguments{
\item{obj}{\link{mosaik}\cr the mosaik (with patches) to modify.}

\item{background}{\link[=integer]{integerish(1)}\cr the value any cell with value
NA should have.}

\item{layer}{\link[=character]{character(1)}\cr the layer in \code{obj} to use.
Defaults to the first layer.}

\item{add}{\link[=character]{character(1)}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\value{
A mosaik of the same dimension as \code{obj}, in which the centroid
of each foreground patch has the value of the patch number and all other
cells have the value \code{background}.
}
\description{
The centroid is the average location of all cells of a foreground patch.
}
\examples{
forest <- mdf_binarise(landscape, match = 41, layer = "cover")
forest <- mdf_componentise(forest)
mdf_centroid(forest)
}
\seealso{
Other operators to determine objects: 
\code{\link{mdf_componentise}()},
\code{\link{mdf_skeletonise}()}
}
\concept{operators to determine objects}
