% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_dilate.R
\name{mdf_dilate}
\alias{mdf_dilate}
\title{Morphologically dilate a mosaik}
\usage{
mdf_dilate(obj = NULL, struct = NULL, layer = NULL, add = NULL)
}
\arguments{
\item{obj}{\code{\link{mosaik}}\cr the mosaik to modify.}

\item{struct}{\code{\link[=struct]{struct(1)}}\cr the structuring element; see
\code{\link{msk_struct}} for details.}

\item{layer}{\code{\link[=character]{character(1)}}\cr the layer in \code{obj} to use.
Defaults to the first layer.}

\item{add}{\code{\link[=character]{character(1)}}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\value{
A mosaik of the same dimension as \code{obj}, where a dilation has
been performed.
}
\description{
The morphological operation 'dilate' increases the value of cells that match
a structuring element.
}
\examples{
# the forest grown by one cell, and by two cells with a 5 x 5 square
m <- landscape |>
  mdf_binarise(match = 47, layer = "cover", add = "forest") |>
  mdf_dilate(layer = "forest", add = "dilated") |>
  mdf_dilate(struct = msk_struct("square", width = 5), layer = "forest",
             add = "dilated_5")
msk_vis(m, .layer("forest"), .layer("dilated"), .layer("dilated_5"))
}
\seealso{
Other operators to morphologically modify a raster: 
\code{\link{mdf_erode}()},
\code{\link{mdf_fill}()},
\code{\link{mdf_morph}()}
}
\concept{operators to morphologically modify a raster}
