% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_mask.R
\name{mdf_mask}
\alias{mdf_mask}
\title{Select cells based on a mask}
\usage{
mdf_mask(obj = NULL, by, background = NA, layer = NULL, add = NULL)
}
\arguments{
\item{obj}{\link{mosaik}\cr the mosaik to modify.}

\item{by}{\link[=character]{character(1)}\cr name of a binary layer in \code{obj}
where cells to retain have value 1 and all others value 0. To mask by a
layer held in another mosaik, bring it in with \code{\link{msk_add}} first.}

\item{background}{\link[=integer]{integerish(1)}\cr the value any masked cell
should have (default \code{NA}).}

\item{layer}{\link[=character]{character(1)}\cr the layer in \code{obj} to use.
Defaults to the first layer.}

\item{add}{\link[=character]{character(1)}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\value{
A mosaik of the same dimensions as \code{obj}, where masked cells
have been set to \code{background}.
}
\description{
Transform a mosaik by setting all cells that are not covered by a mask to
\code{background}.
}
\examples{
# mask one layer by a binary layer of the same mosaik
m <- mdf_binarise(landscape, match = 47, layer = "cover", add = "forest")
mdf_mask(m, by = "forest", layer = "intensity")
mdf_mask(m, by = "forest", layer = "intensity", background = 0)
}
\seealso{
Other operators to select a subset of cells: 
\code{\link{mdf_match}()},
\code{\link{mdf_range}()},
\code{\link{mdf_zonify}()}
}
\concept{operators to select a subset of cells}
