% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_perturb.R
\name{mdf_perturb}
\alias{mdf_perturb}
\title{Add random noise to cell values}
\usage{
mdf_perturb(obj = NULL, sd = 1, layer = NULL, add = NULL)
}
\arguments{
\item{obj}{\code{\link{mosaik}}\cr the mosaik to modify.}

\item{sd}{\code{\link[=numeric]{numeric(1)}}\cr standard deviation of the noise.
Defaults to 1.}

\item{layer}{\code{\link[=character]{character(1)}}\cr the layer to perturb. Defaults to
the first layer.}

\item{add}{\code{\link[=character]{character(1)}}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\value{
A mosaik of the same dimensions with perturbed cell values. NA cells
remain NA.
}
\description{
Perturb cell values by adding random noise drawn from a normal distribution
with mean 0 and standard deviation \code{sd}.
}
\examples{
# the canopy height with normal noise of standard deviation 2 added
m <- mdf_perturb(landscape, sd = 2, layer = "canopy", add = "perturbed")
msk_vis(m, .layer("canopy"), .layer("perturbed"))
}
\seealso{
Other operators to modify cell values: 
\code{\link{mdf_categorise}()},
\code{\link{mdf_distance}()},
\code{\link{mdf_filter}()},
\code{\link{mdf_interpolate}()},
\code{\link{mdf_loop}()},
\code{\link{mdf_offset}()},
\code{\link{mdf_pad}()},
\code{\link{mdf_permute}()},
\code{\link{mdf_replace}()},
\code{\link{mdf_rotate}()},
\code{\link{mdf_scale}()},
\code{\link{mdf_summarise}()},
\code{\link{mdf_tesselate}()},
\code{\link{mdf_transpose}()}
}
\concept{operators to modify cell values}
