% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_range.R
\name{mdf_range}
\alias{mdf_range}
\title{Select cells based on a value range}
\usage{
mdf_range(
  obj = NULL,
  lower = NULL,
  upper = NULL,
  background = NA,
  layer = NULL,
  add = NULL
)
}
\arguments{
\item{obj}{\link{mosaik}\cr the mosaik to modify.}

\item{lower}{\link[=numeric]{numeric(1)}\cr minimum value above which values
will be selected.}

\item{upper}{\link[=numeric]{numeric(1)}\cr maximum value below which values
will be selected.}

\item{background}{\link[=integer]{integerish(1)}\cr the value any cell outside the
range should have.}

\item{layer}{\link[=character]{character(1)}\cr the layer in \code{obj} to use.
Defaults to the first layer.}

\item{add}{\link[=character]{character(1)}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\value{
A mosaik of the same dimension as \code{obj}, where cells within
the range retain their value and others are set to \code{background}.
}
\description{
Transform a mosaik by setting all cells outside a lower and upper threshold
to \code{background}.
}
\examples{
mdf_range(landscape, lower = 30, layer = "intensity")
mdf_range(landscape, upper = 70, layer = "intensity")
mdf_range(landscape, lower = 30, upper = 70, layer = "intensity")
}
\seealso{
Other operators to select a subset of cells: 
\code{\link{mdf_mask}()},
\code{\link{mdf_match}()},
\code{\link{mdf_zonify}()}
}
\concept{operators to select a subset of cells}
