% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_scale.R
\name{mdf_scale}
\alias{mdf_scale}
\title{Change the scale of values in a mosaik}
\usage{
mdf_scale(obj = NULL, range = NULL, layer = NULL, add = NULL)
}
\arguments{
\item{obj}{\code{\link{mosaik}}\cr the mosaik to modify.}

\item{range}{\code{\link[=numeric]{numeric(2)}}\cr vector of minimum and maximum value
to which the values shall be scaled.}

\item{layer}{\code{\link[=character]{character(1)}}\cr the layer in \code{obj} to use.
Defaults to the first layer.}

\item{add}{\code{\link[=character]{character(1)}}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\value{
A mosaik of the same dimension as \code{obj}, where cell values
have been rescaled.
}
\description{
Change minimum and maximum value and scale all values to that range.
}
\examples{
# the canopy height scaled to 0 to 1, and to -1 to 1
m <- landscape |>
  mdf_scale(range = c(0, 1), layer = "canopy", add = "zero_one") |>
  mdf_scale(range = c(-1, 1), layer = "canopy", add = "minus_one_one")
msk_vis(m, .layer("canopy"), .layer("zero_one"), .layer("minus_one_one"))
}
\seealso{
Other operators to modify cell values: 
\code{\link{mdf_binarise}()},
\code{\link{mdf_categorise}()},
\code{\link{mdf_distance}()},
\code{\link{mdf_filter}()},
\code{\link{mdf_interpolate}()},
\code{\link{mdf_loop}()},
\code{\link{mdf_offset}()},
\code{\link{mdf_pad}()},
\code{\link{mdf_permute}()},
\code{\link{mdf_perturb}()},
\code{\link{mdf_replace}()},
\code{\link{mdf_rotate}()},
\code{\link{mdf_summarise}()},
\code{\link{mdf_tesselate}()},
\code{\link{mdf_transpose}()}
}
\concept{operators to modify cell values}
