% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_utils.R
\name{msk_terra}
\alias{msk_terra}
\title{Convert a mosaik to a SpatRaster}
\usage{
msk_terra(obj)
}
\arguments{
\item{obj}{\code{\link{mosaik}}\cr the mosaik to convert.}
}
\value{
A \code{SpatRaster} with one layer per mosaik layer.
}
\description{
Requires the \pkg{terra} package (listed in Suggests).
}
\examples{
\dontrun{
m <- mosaik(extent = c(0, 10, 0, 10), res = 1,
            vals = list(cover = sample(1:5, 100, replace = TRUE)))
r <- msk_terra(m)
}
}
