% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reduceCpp}
\alias{reduceCpp}
\title{Reduce a list of value vectors (c++)}
\usage{
reduceCpp(lVals, f)
}
\arguments{
\item{lVals}{\link[=list]{list(numeric)}\cr list of numeric vectors with equal
length.}

\item{f}{\link[=function]{function(1)}\cr binary function to combine the vectors.}
}
\value{
a single numeric vector of the same length as the input vectors.
}
\description{
C++ function that uses a binary function to successively combine a list
of flat value vectors (all of equal length).
}
