% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaleMatrixCpp}
\alias{scaleMatrixCpp}
\title{Scale values in a matrix (c++)}
\usage{
scaleMatrixCpp(mat, range)
}
\arguments{
\item{mat}{\link[=matrix]{matrix(numeric)}\cr object in which the values are
scaled.}

\item{range}{\link[=numeric]{numeric(2)}\cr two values (minimum and maximum)
between which the values in \code{mat} shall be scaled.}
}
\value{
a numeric matrix
}
\description{
C++ function that scales the values in a matrix between a provided range
}
\seealso{
Other matrix modify functions: 
\code{\link{diamondSquareCpp}()}
}
\concept{matrix modify functions}
