% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_show.R
\name{show,mosaik-method}
\alias{show,mosaik-method}
\title{Print mosaik in the console}
\usage{
\S4method{show}{mosaik}(object)
}
\arguments{
\item{object}{\link{mosaik}\cr object to \code{show}.}
}
\description{
Print mosaik in the console
}
