% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{show,struct-method}
\alias{show,struct-method}
\title{Print the structuring element}
\usage{
\S4method{show}{struct}(object)
}
\arguments{
\item{object}{\link{struct}\cr the structuring element to print.}
}
\description{
Print the structuring element
}
