% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_accessors.R
\name{accessors}
\alias{accessors}
\alias{msk_extent}
\alias{msk_dims}
\alias{msk_res}
\alias{msk_ncells}
\alias{msk_crs}
\alias{msk_provenance}
\alias{msk_names}
\alias{msk_categories}
\alias{msk_patches}
\alias{msk_global}
\title{Accessor functions for mosaik objects}
\usage{
msk_extent(x)

msk_dims(x)

msk_res(x)

msk_ncells(x)

msk_crs(x)

msk_provenance(x)

msk_names(x)

msk_categories(x)

msk_patches(x)

msk_global(x)
}
\arguments{
\item{x}{a \code{\link{mosaik}} object.}
}
\description{
Simple functions to access properties of a \code{\link{mosaik}} object.
}
\examples{
msk <- mosaik(extent = c(0, 10, 0, 10), res = 1,
              vals = list(cover = sample(1:5, 100, replace = TRUE)))
msk_extent(msk)
msk_dims(msk)
msk_res(msk)
msk_ncells(msk)
msk_crs(msk)
msk_names(msk)
}
