% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_binarise.R
\name{mdf_binarise}
\alias{mdf_binarise}
\title{Binarise a mosaik}
\usage{
mdf_binarise(obj = NULL, thresh = NULL, match = NULL, layer = NULL, add = NULL)
}
\arguments{
\item{obj}{\link{mosaik}\cr the mosaik to modify.}

\item{thresh}{\link[=numeric]{numeric(1)}\cr value above which the cell will be
set to 1, below which it will be set to 0.}

\item{match}{\link[=numeric]{numeric(.)}\cr one or more values which will be set
to 1, while the remaining values will be set to 0.}

\item{layer}{\link[=character]{character(1)}\cr the layer in \code{obj} to use.
Defaults to the first layer.}

\item{add}{\link[=character]{character(1)}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\value{
A mosaik of the same dimension as \code{obj}.
}
\description{
Transform a mosaik so that it has the values 0 and 1.
}
\examples{
mdf_binarise(landscape, thresh = 50, layer = "intensity")
mdf_binarise(landscape, match = c(31, 41, 44), layer = "cover")
}
\seealso{
Other operators to modify cell values: 
\code{\link{mdf_categorise}()},
\code{\link{mdf_distance}()},
\code{\link{mdf_filter}()},
\code{\link{mdf_interpolate}()},
\code{\link{mdf_loop}()},
\code{\link{mdf_offset}()},
\code{\link{mdf_pad}()},
\code{\link{mdf_permute}()},
\code{\link{mdf_perturb}()},
\code{\link{mdf_replace}()},
\code{\link{mdf_rotate}()},
\code{\link{mdf_scale}()},
\code{\link{mdf_tesselate}()},
\code{\link{mdf_transpose}()},
\code{\link{mdf_zonal}()}
}
\concept{operators to modify cell values}
