% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_categorise.R
\name{mdf_categorise}
\alias{mdf_categorise}
\title{Categorise a mosaik}
\usage{
mdf_categorise(obj = NULL, breaks = NULL, n = NULL, layer = NULL, add = NULL)
}
\arguments{
\item{obj}{\link{mosaik}\cr the mosaik to modify.}

\item{breaks}{\link[=numeric]{numeric(.)}\cr the break points where categories
should be delimited.}

\item{n}{\link[=integer]{integerish(1)}\cr number of categories.}

\item{layer}{\link[=character]{character(1)}\cr the layer in \code{obj} to use.
Defaults to the first layer.}

\item{add}{\link[=character]{character(1)}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\value{
A mosaik of the same dimension as \code{obj}, in which cells have
the category value into which their values fall.
}
\description{
Transform the values of a mosaik to a (smaller) set of categories.
}
\details{
Using \code{n} will determine \code{breaks} based on the value
range of \code{obj} so that values are assigned to n categories.
}
\examples{
mdf_categorise(landscape, n = 3, layer = "intensity")
mdf_categorise(landscape, breaks = c(25, 50, 75), layer = "intensity")
}
\seealso{
Other operators to modify cell values: 
\code{\link{mdf_binarise}()},
\code{\link{mdf_distance}()},
\code{\link{mdf_filter}()},
\code{\link{mdf_interpolate}()},
\code{\link{mdf_loop}()},
\code{\link{mdf_offset}()},
\code{\link{mdf_pad}()},
\code{\link{mdf_permute}()},
\code{\link{mdf_perturb}()},
\code{\link{mdf_replace}()},
\code{\link{mdf_rotate}()},
\code{\link{mdf_scale}()},
\code{\link{mdf_tesselate}()},
\code{\link{mdf_transpose}()},
\code{\link{mdf_zonal}()}
}
\concept{operators to modify cell values}
