% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_componentise.R
\name{mdf_componentise}
\alias{mdf_componentise}
\title{Determine patches in a mosaik}
\usage{
mdf_componentise(
  obj = NULL,
  connectivity = 4L,
  background = NA,
  layer = NULL,
  add = NULL
)
}
\arguments{
\item{obj}{\link{mosaik}\cr a binarised mosaik to modify.}

\item{connectivity}{\link[=integer]{integer(1)}\cr neighbourhood connectivity;
\code{4} for rook/diamond (default) or \code{8} for queen/box.}

\item{background}{\link[=integer]{integerish(1)}\cr the value any cell with value
NA should have.}

\item{layer}{\link[=character]{character(1)}\cr the layer in \code{obj} to use.
Defaults to the first layer.}

\item{add}{\link[=character]{character(1)}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\value{
A mosaik of the same dimension as \code{obj}, in which neighboring
cells of the foreground have been assigned the same value, forming patches.
}
\description{
Patches are sets of cells that are connected.
}
\examples{
forest <- mdf_binarise(landscape, match = 47, layer = "cover")
mdf_componentise(forest)
mdf_componentise(forest, connectivity = 8L)
}
\seealso{
Other operators to determine objects: 
\code{\link{mdf_centroid}()},
\code{\link{mdf_skeletonise}()}
}
\concept{operators to determine objects}
