% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_erode.R
\name{mdf_erode}
\alias{mdf_erode}
\title{Morphologically erode a mosaik}
\usage{
mdf_erode(obj = NULL, struct = NULL, layer = NULL, add = NULL)
}
\arguments{
\item{obj}{\link{mosaik}\cr the mosaik to modify.}

\item{struct}{\link[=struct]{struct(1)}\cr the structuring element; see
\code{\link{msk_struct}} for details.}

\item{layer}{\link[=character]{character(1)}\cr the layer in \code{obj} to use.
Defaults to the first layer.}

\item{add}{\link[=character]{character(1)}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\value{
A mosaik of the same dimension as \code{obj}, where an erosion has
been performed.
}
\description{
The morphological operation 'erode' decreases the value of cells that match a
structuring element.
}
\examples{
forest <- mdf_binarise(landscape, match = 47, layer = "cover")
mdf_erode(forest)
mdf_erode(forest, struct = msk_struct("square", width = 3))
}
\seealso{
Other operators to morphologically modify a raster: 
\code{\link{mdf_dilate}()},
\code{\link{mdf_fill}()},
\code{\link{mdf_morph}()}
}
\concept{operators to morphologically modify a raster}
