% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_match.R
\name{mdf_match}
\alias{mdf_match}
\title{Select cells based on a structuring element}
\usage{
mdf_match(
  obj = NULL,
  struct = NULL,
  rotate = TRUE,
  background = NA,
  layer = NULL,
  add = NULL
)
}
\arguments{
\item{obj}{\link{mosaik}\cr the mosaik to modify.}

\item{struct}{\link[=struct]{struct(1)}\cr the pattern to match; see
\code{\link{msk_struct}} for details.}

\item{rotate}{\link[=logical]{logical(1)}\cr whether the kernel should be rotated
for all possible rotations or used as is.}

\item{background}{\link[=integer]{integerish(1)}\cr the value any non-matching cell
should have.}

\item{layer}{\link[=character]{character(1)}\cr the layer in \code{obj} to use.
Defaults to the first layer.}

\item{add}{\link[=character]{character(1)}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\value{
A mosaik of the same dimension as \code{obj}.
}
\description{
Transform a mosaik by setting all cells that do not match a structuring
element to \code{background}. Also known as the 'hit-or-miss'-transform.
}
\examples{
forest <- mdf_binarise(landscape, match = 47, layer = "cover")
s <- msk_struct("cross", width = 3)
mdf_match(forest, struct = s)
mdf_match(forest, struct = s, rotate = FALSE)
}
\seealso{
Other operators to select a subset of cells: 
\code{\link{mdf_mask}()},
\code{\link{mdf_range}()},
\code{\link{mdf_zonify}()}
}
\concept{operators to select a subset of cells}
