% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_morph.R
\name{mdf_morph}
\alias{mdf_morph}
\title{Morphologically modify a mosaik}
\usage{
mdf_morph(
  obj = NULL,
  struct = NULL,
  blend = NULL,
  merge = NULL,
  rotate = TRUE,
  strict = TRUE,
  background = NA,
  layer = NULL,
  add = NULL
)
}
\arguments{
\item{obj}{\link{mosaik}\cr the mosaik to modify.}

\item{struct}{\link[=struct]{struct(1)}\cr the structuring element; see
\code{\link{msk_struct}} for details.}

\item{blend}{\link[=character]{character(1)}\cr \code{identity}, \code{equal},
\code{lower}, \code{greater}, \code{plus}, \code{minus}, \code{product};
see Details.}

\item{merge}{\link[=character]{character(1)}\cr \code{min}, \code{max}, \code{all},
\code{any}, \code{sum}, \code{mean}, \code{median}, \code{sd}, \code{cv},
\code{one}, \code{zero}, \code{na}; see Details.}

\item{rotate}{\link[=logical]{logical(1)}\cr whether the kernel should be rotated.}

\item{strict}{\link[=logical]{logical(1)}\cr whether ...}

\item{background}{\link[=integer]{integerish(1)}\cr the value any cell with value
NA should have.}

\item{layer}{\link[=character]{character(1)}\cr the layer in \code{obj} to use.
Defaults to the first layer.}

\item{add}{\link[=character]{character(1)}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\description{
Morphologically modify a mosaik
}
\details{
The \code{morphCpp} function (internal) iteratively goes through
each pixel and compares a structuring element with the mosaik at that
location. The result depends on \code{blend} and \code{merge}:
\itemize{
\item First, values covered by the structuring element are blended
pairwise with the element (\code{blend}).
\item Then these values are merged into a single value (\code{merge})
and assigned in the current location.
}

Blend functions: identity, equal, lower, greater, plus, minus, product.
Merge functions: min, max, all, any, !all, !any, sum, mean, median, sd,
cv, sumNa.
}
\examples{
s <- msk_struct("disc", width = 3)

# smoothing: weighted mean of neighbourhood
mdf_morph(landscape, struct = s, blend = "product", merge = "mean",
          layer = "intensity")

# edge detection on binary layer
forest <- mdf_binarise(landscape, match = 47, layer = "cover")
mdf_morph(forest, struct = s, blend = "equal", merge = "all",
          strict = TRUE, background = 0)

# local maximum
mdf_morph(landscape, struct = s, blend = "identity", merge = "max",
          layer = "intensity")
}
\references{
Credit for the original idea and C++ code is due to Jon Clayden
(\href{https://github.com/jonclayden/mmand}{R::mmand}).
}
\seealso{
Other operators to morphologically modify a raster: 
\code{\link{mdf_dilate}()},
\code{\link{mdf_erode}()},
\code{\link{mdf_fill}()}
}
\concept{operators to morphologically modify a raster}
