% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_offset.R
\name{mdf_offset}
\alias{mdf_offset}
\title{Offset the values in a mosaik}
\usage{
mdf_offset(obj = NULL, fun = "+", value = 1, layer = NULL, add = NULL)
}
\arguments{
\item{obj}{\code{\link{mosaik}}\cr the mosaik to modify.}

\item{fun}{\code{\link[=character]{character(1)}}\cr function used for offsetting; supported
are all \link[=^]{arithmetic operators}.}

\item{value}{\code{\link[=numeric]{numeric(1)}}\cr the value by which to offset all
values (can also be negative).}

\item{layer}{\code{\link[=character]{character(1)}}\cr the layer in \code{obj} to use.
Defaults to the first layer.}

\item{add}{\code{\link[=character]{character(1)}}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\value{
A mosaik of the same dimension as \code{obj}.
}
\description{
Apply arithmetic operators to offset all cell values.
}
\examples{
# the canopy height plus 5 and times 2, and the land cover modulo 3
m <- landscape |>
  mdf_offset(fun = "+", value = 5, layer = "canopy", add = "plus_5") |>
  mdf_offset(fun = "*", value = 2, layer = "canopy", add = "times_2") |>
  mdf_offset(fun = "\%\%", value = 3, layer = "cover", add = "modulo_3")
msk_vis(m, .layer("canopy"), .layer("plus_5"), .layer("times_2"))
msk_vis(m, .layer("cover"), .layer("modulo_3"))
}
\seealso{
Other operators to modify cell values: 
\code{\link{mdf_categorise}()},
\code{\link{mdf_distance}()},
\code{\link{mdf_filter}()},
\code{\link{mdf_interpolate}()},
\code{\link{mdf_loop}()},
\code{\link{mdf_pad}()},
\code{\link{mdf_permute}()},
\code{\link{mdf_perturb}()},
\code{\link{mdf_replace}()},
\code{\link{mdf_rotate}()},
\code{\link{mdf_scale}()},
\code{\link{mdf_summarise}()},
\code{\link{mdf_tesselate}()},
\code{\link{mdf_transpose}()}
}
\concept{operators to modify cell values}
