% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_offset.R
\name{mdf_offset}
\alias{mdf_offset}
\title{Offset the values in a mosaik}
\usage{
mdf_offset(obj = NULL, fun = "+", value = 1, layer = NULL, add = NULL)
}
\arguments{
\item{obj}{\link{mosaik}\cr the mosaik to modify.}

\item{fun}{\link[=character]{character(1)}\cr function used for offsetting; supported
are all \link[=^]{arithmetic operators}.}

\item{value}{\link[=numeric]{numeric(1)}\cr the value by which to offset all
values (can also be negative).}

\item{layer}{\link[=character]{character(1)}\cr the layer in \code{obj} to use.
Defaults to the first layer.}

\item{add}{\link[=character]{character(1)}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\value{
A mosaik of the same dimension as \code{obj}.
}
\description{
Apply arithmetic operators to offset all cell values.
}
\examples{
mdf_offset(landscape, fun = "+", value = 10, layer = "intensity")
mdf_offset(landscape, fun = "*", value = 2, layer = "intensity")
mdf_offset(landscape, fun = "\%\%", value = 3, layer = "cover")
}
\seealso{
Other operators to modify cell values: 
\code{\link{mdf_binarise}()},
\code{\link{mdf_categorise}()},
\code{\link{mdf_distance}()},
\code{\link{mdf_filter}()},
\code{\link{mdf_interpolate}()},
\code{\link{mdf_loop}()},
\code{\link{mdf_pad}()},
\code{\link{mdf_permute}()},
\code{\link{mdf_perturb}()},
\code{\link{mdf_replace}()},
\code{\link{mdf_rotate}()},
\code{\link{mdf_scale}()},
\code{\link{mdf_tesselate}()},
\code{\link{mdf_transpose}()},
\code{\link{mdf_zonal}()}
}
\concept{operators to modify cell values}
