% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_permute.R
\name{mdf_permute}
\alias{mdf_permute}
\title{Permute the values in a mosaik}
\usage{
mdf_permute(obj = NULL, type = "invert", by = NULL, layer = NULL, add = NULL)
}
\arguments{
\item{obj}{\link{mosaik}\cr the mosaik to modify.}

\item{type}{\link[=character]{character(1)}\cr the permutation type. Either
\code{"invert"}, \code{"revert"}, \code{"descending"}, \code{"ascending"}
or \code{"cycle"}.}

\item{by}{\link[=integer]{integerish(1)}\cr value by which to apply the
permutation; only for \code{type = "cycle"}.}

\item{layer}{\link[=character]{character(1)}\cr the layer in \code{obj} to use.
Defaults to the first layer.}

\item{add}{\link[=character]{character(1)}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\value{
A mosaik of the same dimensions as \code{obj}.
}
\description{
The permutation of a set of cell values leads to a systematic
(re)arrangement of all the members of the set.
}
\examples{
mdf_permute(landscape, type = "invert", layer = "cover")
mdf_permute(landscape, type = "revert", layer = "cover")
mdf_permute(landscape, type = "ascending", layer = "intensity")
mdf_permute(landscape, type = "descending", layer = "intensity")
mdf_permute(landscape, type = "cycle", by = 2L, layer = "cover")
}
\seealso{
Other operators to modify cell values: 
\code{\link{mdf_binarise}()},
\code{\link{mdf_categorise}()},
\code{\link{mdf_distance}()},
\code{\link{mdf_filter}()},
\code{\link{mdf_interpolate}()},
\code{\link{mdf_loop}()},
\code{\link{mdf_offset}()},
\code{\link{mdf_pad}()},
\code{\link{mdf_perturb}()},
\code{\link{mdf_replace}()},
\code{\link{mdf_rotate}()},
\code{\link{mdf_scale}()},
\code{\link{mdf_tesselate}()},
\code{\link{mdf_transpose}()},
\code{\link{mdf_zonal}()}
}
\concept{operators to modify cell values}
