% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_range.R
\name{mdf_range}
\alias{mdf_range}
\title{Select cells based on a value range}
\usage{
mdf_range(
  obj = NULL,
  lower = NULL,
  upper = NULL,
  background = NA,
  layer = NULL,
  add = NULL
)
}
\arguments{
\item{obj}{\code{\link{mosaik}}\cr the mosaik to modify.}

\item{lower}{\code{\link[=numeric]{numeric(1)}}\cr minimum value above which values
will be selected.}

\item{upper}{\code{\link[=numeric]{numeric(1)}}\cr maximum value below which values
will be selected.}

\item{background}{\code{\link[=integer]{integerish(1)}}\cr the value any cell outside the
range should have.}

\item{layer}{\code{\link[=character]{character(1)}}\cr the layer in \code{obj} to use.
Defaults to the first layer.}

\item{add}{\code{\link[=character]{character(1)}}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\value{
A mosaik of the same dimension as \code{obj}, where cells within
the range retain their value and others are set to \code{background}.
}
\description{
Transform a mosaik by setting all cells outside a lower and upper threshold
to \code{background}.
}
\examples{
# the canopy height from 5 m, up to 20 m, and between the two
m <- landscape |>
  mdf_range(lower = 5, layer = "canopy", add = "from_5") |>
  mdf_range(upper = 20, layer = "canopy", add = "to_20") |>
  mdf_range(lower = 5, upper = 20, layer = "canopy", add = "5_to_20")
msk_vis(m, .layer("canopy"), .layer("from_5"), .layer("to_20"),
        .layer("5_to_20"))
}
\seealso{
Other operators to select a subset of cells: 
\code{\link{mdf_match}()}
}
\concept{operators to select a subset of cells}
