% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_replace.R
\name{mdf_replace}
\alias{mdf_replace}
\title{Replace values in a mosaik}
\usage{
mdf_replace(obj = NULL, old = NULL, new = NULL, layer = NULL, add = NULL)
}
\arguments{
\item{obj}{\link{mosaik}\cr the mosaik to modify.}

\item{old}{\link[=numeric]{numeric(.)}\cr values to be substituted.}

\item{new}{\link[=numeric]{numeric(.)}\cr values to substitute with.}

\item{layer}{\link[=character]{character(1)}\cr the layer in \code{obj} to use.
Defaults to the first layer.}

\item{add}{\link[=character]{character(1)}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\value{
A mosaik of the same dimensions as \code{obj}.
}
\description{
Replace a set of values with new values.
}
\examples{
mdf_replace(landscape, old = 47, new = 99, layer = "cover")
mdf_replace(landscape, old = c(21, 24), new = c(100, 200),
            layer = "cover")
}
\seealso{
Other operators to modify cell values: 
\code{\link{mdf_binarise}()},
\code{\link{mdf_categorise}()},
\code{\link{mdf_distance}()},
\code{\link{mdf_filter}()},
\code{\link{mdf_interpolate}()},
\code{\link{mdf_loop}()},
\code{\link{mdf_offset}()},
\code{\link{mdf_pad}()},
\code{\link{mdf_permute}()},
\code{\link{mdf_perturb}()},
\code{\link{mdf_rotate}()},
\code{\link{mdf_scale}()},
\code{\link{mdf_tesselate}()},
\code{\link{mdf_transpose}()},
\code{\link{mdf_zonal}()}
}
\concept{operators to modify cell values}
