% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_replace.R
\name{mdf_replace}
\alias{mdf_replace}
\title{Replace values in a mosaik}
\usage{
mdf_replace(obj = NULL, old = NULL, new = NULL, layer = NULL, add = NULL)
}
\arguments{
\item{obj}{\code{\link{mosaik}}\cr the mosaik to modify.}

\item{old}{\code{\link[=numeric]{numeric(.)}}\cr values to be substituted.}

\item{new}{\code{\link[=numeric]{numeric(.)}}\cr values to substitute with.}

\item{layer}{\code{\link[=character]{character(1)}}\cr the layer in \code{obj} to use.
Defaults to the first layer.}

\item{add}{\code{\link[=character]{character(1)}}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\value{
A mosaik of the same dimensions as \code{obj}.
}
\description{
Replace a set of values with new values.
}
\examples{
# class 47 renumbered to 99, and classes 21 and 24 to 100 and 200
m <- landscape |>
  mdf_replace(old = 47, new = 99, layer = "cover", add = "one") |>
  mdf_replace(old = c(21, 24), new = c(100, 200), layer = "cover",
              add = "two")
msk_vis(m, .layer("cover"), .layer("one"), .layer("two"))
}
\seealso{
Other operators to modify cell values: 
\code{\link{mdf_categorise}()},
\code{\link{mdf_distance}()},
\code{\link{mdf_filter}()},
\code{\link{mdf_interpolate}()},
\code{\link{mdf_loop}()},
\code{\link{mdf_offset}()},
\code{\link{mdf_pad}()},
\code{\link{mdf_permute}()},
\code{\link{mdf_perturb}()},
\code{\link{mdf_rotate}()},
\code{\link{mdf_scale}()},
\code{\link{mdf_summarise}()},
\code{\link{mdf_tesselate}()},
\code{\link{mdf_transpose}()}
}
\concept{operators to modify cell values}
