% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_resize.R
\name{mdf_resize}
\alias{mdf_resize}
\title{Change the size of a mosaik}
\usage{
mdf_resize(obj = NULL, factor = NULL, layer = NULL)
}
\arguments{
\item{obj}{\code{\link{mosaik}}\cr the mosaik to modify.}

\item{factor}{\code{\link[=numeric]{numeric(1)}}\cr an integer larger than 1 for
up-scaling and a fraction for down-scaling (see Details).}

\item{layer}{\code{\link[=character]{character(1)}}\cr the layer in \code{obj} to use.
Defaults to the first layer.}
}
\value{
A mosaik with rescaled dimensions.
}
\description{
Increase or decrease the size of a mosaik using nearest-neighbour
interpolation.
}
\details{
\code{factor} multiplies the number of cells in x and y. If
\code{factor > 1}, the object is up-scaled; if \code{factor < 1}, it is
down-scaled. Allowed values are integers or inverse integers (1/n).
}
\examples{
# twice and a quarter the number of cells along each side
m <- mdf_resize(landscape, factor = 2)
msk_vis(m, .layer("cover"))

m <- mdf_resize(landscape, factor = 1/4)
msk_vis(m, .layer("cover"))
}
\seealso{
Other operators to modify the overall object: 
\code{\link{mdf_blend}()}
}
\concept{operators to modify the overall object}
