% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_rotate.R
\name{mdf_rotate}
\alias{mdf_rotate}
\title{Rotate a mosaik by 90-degree increments}
\usage{
mdf_rotate(obj = NULL, angle = 90L, layer = NULL)
}
\arguments{
\item{obj}{\code{\link{mosaik}}\cr the mosaik to modify.}

\item{angle}{\code{\link[=integer]{integerish(1)}}\cr rotation angle in degrees. Must be
one of \code{90}, \code{180}, or \code{270}.}

\item{layer}{\code{\link[=character]{character(1)}}\cr the layer to rotate. If
\code{NULL} (default), all layers are rotated.}
}
\value{
A mosaik with rotated cell values and adjusted extent/dimensions.
}
\description{
Rotate the grid clockwise by 90, 180, or 270 degrees. Extent and dimensions
are adjusted accordingly.
}
\examples{
# the map turned by 90, 180 and 270 degrees
msk_vis(mdf_rotate(landscape, angle = 90), .layer("cover"))
msk_vis(mdf_rotate(landscape, angle = 180), .layer("cover"))
msk_vis(mdf_rotate(landscape, angle = 270), .layer("cover"))
}
\seealso{
Other operators to modify cell values: 
\code{\link{mdf_binarise}()},
\code{\link{mdf_categorise}()},
\code{\link{mdf_distance}()},
\code{\link{mdf_filter}()},
\code{\link{mdf_interpolate}()},
\code{\link{mdf_loop}()},
\code{\link{mdf_offset}()},
\code{\link{mdf_pad}()},
\code{\link{mdf_permute}()},
\code{\link{mdf_perturb}()},
\code{\link{mdf_replace}()},
\code{\link{mdf_scale}()},
\code{\link{mdf_summarise}()},
\code{\link{mdf_tesselate}()},
\code{\link{mdf_transpose}()}
}
\concept{operators to modify cell values}
