% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdf_scale.R
\name{mdf_scale}
\alias{mdf_scale}
\title{Change the scale of values in a mosaik}
\usage{
mdf_scale(obj = NULL, range = NULL, layer = NULL, add = NULL)
}
\arguments{
\item{obj}{\link{mosaik}\cr the mosaik to modify.}

\item{range}{\link[=numeric]{numeric(2)}\cr vector of minimum and maximum value
to which the values shall be scaled.}

\item{layer}{\link[=character]{character(1)}\cr the layer in \code{obj} to use.
Defaults to the first layer.}

\item{add}{\link[=character]{character(1)}\cr if \code{NULL} (default), overwrite
\code{layer}; if a string, write to a new layer with that name.}
}
\value{
A mosaik of the same dimension as \code{obj}, where cell values
have been rescaled.
}
\description{
Change minimum and maximum value and scale all values to that range.
}
\examples{
mdf_scale(landscape, range = c(0, 1), layer = "intensity")
mdf_scale(landscape, range = c(-1, 1), layer = "intensity")
}
\seealso{
Other operators to modify cell values: 
\code{\link{mdf_binarise}()},
\code{\link{mdf_categorise}()},
\code{\link{mdf_distance}()},
\code{\link{mdf_filter}()},
\code{\link{mdf_interpolate}()},
\code{\link{mdf_loop}()},
\code{\link{mdf_offset}()},
\code{\link{mdf_pad}()},
\code{\link{mdf_permute}()},
\code{\link{mdf_perturb}()},
\code{\link{mdf_replace}()},
\code{\link{mdf_rotate}()},
\code{\link{mdf_tesselate}()},
\code{\link{mdf_transpose}()},
\code{\link{mdf_zonal}()}
}
\concept{operators to modify cell values}
