% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\docType{class}
\name{struct-class}
\alias{struct-class}
\alias{struct}
\title{The \code{struct} class (S4) and its methods}
\description{
A structuring element is any 2D array with an odd number of cells in each
dimension and a focal cell in the middle. Each cell is compared against a
mosaik to probe the local pattern. Typically a conclusion is drawn based on
how the structuring element matches the local pattern and this conclusion is
stored in the value of the focal cell.
}
\details{
The cells of a structuring element can have four values: NA (indifferent), 0,
1, and any value > 1.
}
\section{Slots}{

\describe{
\item{\code{pattern}}{\link[=matrix]{matrix(.)}\cr the pattern matrix.}

\item{\code{rotate}}{\link[=logical]{logical(1)}\cr whether the pattern shall be applied
in all rotations.}

\item{\code{background}}{\link[=numeric]{numeric(1)}\cr the value of background cells.}
}}

